+++ airsense_luxury-main/airsense_luxury-main/libraries/DFRobot_OzoneSensor-master/library.properties
@@ -1,0 +1,9 @@
+name=DFRobot_OzoneSensor
+version=1.0.0
+author=DFRobot
+maintainer=ZhixinLiu <zhixin.liu@dfrobot.com>
+sentence=DFRobot Standard library.
+paragraph=DFRobot_OzoneSensor is the standard library of DFRobot.
+category=Sensors
+url=https://github.com/DFRobot/DFRobot_OzoneSensor
+architectures=*
