+++ airsense_luxury-main/airsense_luxury-main/libraries/arduino-sht-master/SHTSensor.cpp
@@ -27,7 +27,6 @@
  */
 
 #include <inttypes.h>
-#include <Wire.h>
 #include <Arduino.h>
 
 #include "SHTSensor.h"
@@ -51,37 +50,37 @@ bool SHTSensorDriver::readSample()
 // class SHTI2cSensor
 //
 
-const uint8_t SHTI2cSensor::CMD_SIZE             = 2;
 const uint8_t SHTI2cSensor::EXPECTED_DATA_SIZE   = 6;
 
-bool SHTI2cSensor::readFromI2c(uint8_t i2cAddress,
+bool SHTI2cSensor::readFromI2c(TwoWire & wire,
+                               uint8_t i2cAddress,
                                const uint8_t *i2cCommand,
                                uint8_t commandLength, uint8_t *data,
                                uint8_t dataLength,
                                uint8_t duration)
 {
-  Wire.beginTransmission(i2cAddress);
+  wire.beginTransmission(i2cAddress);
   for (int i = 0; i < commandLength; ++i) {
-    if (Wire.write(i2cCommand[i]) != 1) {
+    if (wire.write(i2cCommand[i]) != 1) {
       return false;
     }
   }
 
-  if (Wire.endTransmission(false) != 0) {
+  if (wire.endTransmission() != 0) {
     return false;
   }
 
   delay(duration);
 
-  Wire.requestFrom(i2cAddress, dataLength);
+  wire.requestFrom(i2cAddress, dataLength);
 
   // check if the same number of bytes are received that are requested.
-  if (Wire.available() != dataLength) {
+  if (wire.available() != dataLength) {
     return false;
   }
 
   for (int i = 0; i < dataLength; ++i) {
-    data[i] = Wire.read();
+    data[i] = wire.read();
   }
   return true;
 }
@@ -110,12 +109,13 @@ uint8_t SHTI2cSensor::crc8(const uint8_t *data, uint8_t len)
 bool SHTI2cSensor::readSample()
 {
   uint8_t data[EXPECTED_DATA_SIZE];
-  uint8_t cmd[CMD_SIZE];
+  uint8_t cmd[mCmd_Size];
 
   cmd[0] = mI2cCommand >> 8;
+  //is omitted for SHT4x Sensors
   cmd[1] = mI2cCommand & 0xff;
 
-  if (!readFromI2c(mI2cAddress, cmd, CMD_SIZE, data,
+  if (!readFromI2c(mWire, mI2cAddress, cmd, mCmd_Size, data,
                    EXPECTED_DATA_SIZE, mDuration)) {
     return false;
   }
@@ -133,11 +133,11 @@ bool SHTI2cSensor::readSample()
   mTemperature = mA + mB * (val / mC);
 
   val = (data[3] << 8) + data[4];
-  mHumidity = mX * (val / mY);
+  mHumidity = mX + mY * (val / mZ);
 
   return true;
-}
 
+}
 
 //
 // class SHTC1Sensor
@@ -146,9 +146,9 @@ bool SHTI2cSensor::readSample()
 class SHTC1Sensor : public SHTI2cSensor
 {
 public:
-    SHTC1Sensor()
+    SHTC1Sensor(TwoWire & wire)
         // clock stretching disabled, high precision, T first
-        : SHTI2cSensor(0x70, 0x7866, 15, -45, 175, 65535, 100, 65535)
+        : SHTI2cSensor(0x70, 0x7866, 15, -45, 175, 65535, 0, 100, 65535, 2, wire)
     {
     }
 };
@@ -173,10 +173,10 @@ public:
   static const uint8_t SHT3X_I2C_ADDRESS_44 = 0x44;
   static const uint8_t SHT3X_I2C_ADDRESS_45 = 0x45;
 
-  SHT3xSensor(uint8_t i2cAddress = SHT3X_I2C_ADDRESS_44)
+  SHT3xSensor(TwoWire & wire, uint8_t i2cAddress = SHT3X_I2C_ADDRESS_44)
       : SHTI2cSensor(i2cAddress, SHT3X_ACCURACY_HIGH,
                      SHT3X_ACCURACY_HIGH_DURATION,
-                     -45, 175, 65535, 100, 65535)
+                     -45, 175, 65535, 0, 100, 65535, 2, wire)
   {
   }
 
@@ -194,6 +194,55 @@ public:
       case SHTSensor::SHT_ACCURACY_LOW:
         mI2cCommand = SHT3X_ACCURACY_LOW;
         mDuration = SHT3X_ACCURACY_LOW_DURATION;
+        break;
+      default:
+        return false;
+    }
+    return true;
+  }
+};
+
+
+//
+// class SHT4xSensor
+//
+
+class SHT4xSensor : public SHTI2cSensor
+{
+private:
+  static const uint16_t SHT4X_ACCURACY_HIGH    = 0xFD00;
+  static const uint16_t SHT4X_ACCURACY_MEDIUM  = 0xF600;
+  static const uint16_t SHT4X_ACCURACY_LOW     = 0xE000;
+
+  static const uint8_t SHT4X_ACCURACY_HIGH_DURATION   = 10;
+  static const uint8_t SHT4X_ACCURACY_MEDIUM_DURATION = 4;
+  static const uint8_t SHT4X_ACCURACY_LOW_DURATION    = 2;
+
+public:
+  static const uint8_t SHT4X_I2C_ADDRESS_44 = 0x44;
+  static const uint8_t SHT4X_I2C_ADDRESS_45 = 0x45;
+
+  SHT4xSensor(TwoWire & wire, uint8_t i2cAddress = SHT4X_I2C_ADDRESS_44)
+      : SHTI2cSensor(i2cAddress, SHT4X_ACCURACY_HIGH,
+                     SHT4X_ACCURACY_HIGH_DURATION,
+                     -45, 175, 65535, -6, 125, 65535, 1, wire)
+  {
+  }
+
+  virtual bool setAccuracy(SHTSensor::SHTAccuracy newAccuracy)
+  {
+    switch (newAccuracy) {
+      case SHTSensor::SHT_ACCURACY_HIGH:
+        mI2cCommand = SHT4X_ACCURACY_HIGH;
+        mDuration = SHT4X_ACCURACY_HIGH_DURATION;
+        break;
+      case SHTSensor::SHT_ACCURACY_MEDIUM:
+        mI2cCommand = SHT4X_ACCURACY_MEDIUM;
+        mDuration = SHT4X_ACCURACY_MEDIUM_DURATION;
+        break;
+      case SHTSensor::SHT_ACCURACY_LOW:
+        mI2cCommand = SHT4X_ACCURACY_LOW;
+        mDuration = SHT4X_ACCURACY_LOW_DURATION;
         break;
       default:
         return false;
@@ -227,12 +276,13 @@ float SHT3xAnalogSensor::readTemperature()
 const SHTSensor::SHTSensorType SHTSensor::AUTO_DETECT_SENSORS[] = {
   SHT3X,
   SHT3X_ALT,
-  SHTC1
+  SHTC1,
+  SHT4X
 };
 const float SHTSensor::TEMPERATURE_INVALID = NAN;
 const float SHTSensor::HUMIDITY_INVALID = NAN;
 
-bool SHTSensor::init()
+bool SHTSensor::init(TwoWire & wire)
 {
   if (mSensor != NULL) {
     cleanup();
@@ -240,19 +290,23 @@ bool SHTSensor::init()
 
   switch(mSensorType) {
     case SHT3X:
-      mSensor = new SHT3xSensor();
+    case SHT85:
+      mSensor = new SHT3xSensor(wire);
       break;
 
     case SHT3X_ALT:
-      mSensor = new SHT3xSensor(SHT3xSensor::SHT3X_I2C_ADDRESS_45);
+      mSensor = new SHT3xSensor(wire, SHT3xSensor::SHT3X_I2C_ADDRESS_45);
       break;
 
     case SHTW1:
     case SHTW2:
     case SHTC1:
-      mSensor = new SHTC1Sensor();
+    case SHTC3:
+      mSensor = new SHTC1Sensor(wire);
       break;
-
+    case SHT4X:
+      mSensor = new SHT4xSensor(wire);
+      break;
     case AUTO_DETECT:
     {
       bool detected = false;
@@ -260,7 +314,8 @@ bool SHTSensor::init()
            i < sizeof(AUTO_DETECT_SENSORS) / sizeof(AUTO_DETECT_SENSORS[0]);
            ++i) {
         mSensorType = AUTO_DETECT_SENSORS[i];
-        if (init() && readSample()) {
+        delay(40); // TODO: this was necessary to make SHT4x autodetect work; revisit to find root cause
+        if (init(wire)) {
           detected = true;
           break;
         }
@@ -271,7 +326,10 @@ bool SHTSensor::init()
       break;
     }
   }
-  return (mSensor != NULL);
+
+  // to finish the initialization, attempt to read to make sure the communication works
+  // Note: readSample() will check for a NULL mSensor in case auto detect failed
+  return readSample();
 }
 
 bool SHTSensor::readSample()
