+++ airsense_luxury-main/airsense_luxury-main/libraries/arduino-sht-master/arduino-sht.h
@@ -1,0 +1,6 @@
+#ifndef ARDUINO_SHT_H
+#define ARDUINO_SHT_H
+
+#include "SHTSensor.h"
+
+#endif
