+++ airsense_luxury-main/airsense_luxury-main/libraries/arduino-sht-master/library.properties
@@ -1,9 +1,9 @@
 name=arduino-sht
-version=1.1.0
+version=1.2.2
 author=Johannes Winkelmann, Andreas Brauchli
-maintainer=Andreas Brauchli <andreas.brauchli@sensirion.com>
+maintainer=Johannes Winkelmann <jwi@sensirion.com>
 sentence=Support for Sensirion's humidity and temperature sensors.
-paragraph=Supported sensors: SHTC1, SHTW1, SHTW2, SHT3x-DIS (I2C), SHT3x-ARP.
+paragraph=Supported sensors: SHTC1, SHTC3, SHTW1, SHTW2, SHT3x-DIS (I2C), SHT85, SHT3x-ARP, SHT4x
 category=Sensors
 url=https://developer.sensirion.com
 architectures=*
